+++ controllers/donation.go
@@ -19,7 +19,6 @@ import (
 	log "github.com/Sirupsen/logrus"
 	"github.com/gin-gonic/gin"
 	"github.com/gin-gonic/gin/binding"
-	"github.com/jinzhu/copier"
 	"gopkg.in/go-playground/validator.v8"
 	"gopkg.in/guregu/null.v3"
 
@@ -29,28 +28,28 @@ import (
 
 type (
 	clientReq struct {
-		Amount      uint              `json:"amount" form:"amount" binding:"required"`
-		Cardholder  models.Cardholder `json:"cardholder" form:"cardholder" binding:"required,dive"`
-		Currency    string            `json:"currency" form:"currency"`
-		Details     string            `json:"details" form:"details"`
-		MerchantID  string            `json:"merchant_id" form:"merchant_id"`
-		OrderNumber string            `json:"order_number" form:"order_number"`
-		Prime       string            `json:"prime" form:"prime" binding:"required"`
-		ResultUrl   linePayResultUrl  `json:"result_url" form:"result_url"`
-		UserID      uint              `json:"user_id" form:"user_id" binding:"required"`
+		Amount     uint              `json:"amount" form:"amount" binding:"required"`
+		Cardholder models.Cardholder `json:"donator" form:"donator" binding:"required,dive"`
+		Currency   string            `json:"currency" form:"currency"`
+		Details    string            `json:"details" form:"details"`
+		MerchantID string            `json:"merchant_id" form:"merchant_id"`
+		Prime      string            `json:"prime" form:"prime" binding:"required"`
+		ResultUrl  linePayResultUrl  `json:"result_url" form:"result_url"`
+		Frequency  string            `json:"frequency"`
+		UserID     uint              `json:"user_id" form:"user_id" binding:"required"`
 	}
 
 	clientResp struct {
 		Amount      uint              `json:"amount"`
-		CardInfo    models.CardInfo   `json:"card_info,omitempty"`
-		Cardholder  models.Cardholder `json:"cardholder,omitempty"`
+		CardInfo    models.CardInfo   `json:"card_info"`
+		Cardholder  models.Cardholder `json:"cardholder"`
 		Currency    string            `json:"currency"`
-		Details     string            `json:"details,omitempty"`
+		Details     string            `json:"details"`
+		Frequency   string            `json:"frequency"`
 		ID          uint              `json:"id"`
 		Notes       string            `json:"notes"`
-		OrderNumber string            `json:"order_number,omitempty"`
+		OrderNumber string            `json:"order_number"`
 		PayMethod   string            `json:"pay_method"`
-		PeriodicID  uint              `json:"periodic_id,omitempty"`
 		SendReceipt string            `json:"send_receipt"`
 		ToFeedback  bool              `json:"to_feedback"`
 	}
@@ -71,34 +70,25 @@ type (
 	}
 
 	tapPayPrimeReq struct {
-		Prime       string            `json:"prime"`
-		PartnerKey  string            `json:"partner_key"`
-		MerchantID  string            `json:"merchant_id"`
 		Amount      uint              `json:"amount"`
+		Cardholder  models.Cardholder `json:"cardholder"`
 		Currency    string            `json:"currency"`
 		Details     string            `json:"details"`
-		Cardholder  models.Cardholder `json:"cardholder"`
-		Remember    bool              `json:"remember"`
+		MerchantID  string            `json:"merchant_id"`
 		OrderNumber string            `json:"order_number"`
+		PartnerKey  string            `json:"partner_key"`
+		Prime       string            `json:"prime"`
+		Remember    bool              `json:"remember"`
 		ResultUrl   linePayResultUrl  `json:"result_url"`
 	}
 
 	tapPayTransactionResp struct {
-		Status                int64               `json:"status"`
-		Msg                   string              `json:"msg"`
-		RecTradeID            string              `json:"rec_trade_id"`
-		BankTransactionID     string              `json:"bank_transaction_id"`
-		AuthCode              string              `json:"auth_code"`
-		CardSecret            cardSecret          `json:"card_secret"`
-		Amount                uint                `json:"amount"`
-		Currency              string              `json:"currency"`
-		CardInfo              models.CardInfo     `json:"card_info"`
-		OrderNumber           string              `json:"order_number"`
-		Acquirer              string              `json:"acquirer"`
-		TransactionTimeMillis int64               `json:"transaction_time_millis"`
+		models.TappayResp
 		BankTransactionTime   bankTransactionTime `json:"bank_transaction_time"`
-		BankResultCode        null.String         `json:"bank_result_code"`
-		BankResultMsg         null.String         `json:"bank_result_msg"`
+		CardInfo              models.CardInfo     `json:"card_info"`
+		CardSecret            cardSecret          `json:"card_secret"`
+		Status                int64               `json:"status"`
+		TransactionTimeMillis int64               `json:"transaction_time_millis"`
 	}
 
 	tapPayMinTransactionResp struct {
@@ -109,44 +99,181 @@ type (
 	payType int
 
 	patchBody struct {
-		Address     null.String `json:"address"`
-		Details     null.String `json:"details"`
-		Name        null.String `json:"name"`
-		NationalID  null.String `json:"national_id"`
-		Notes       null.String `json:"notes"`
-		PhoneNumber null.String `json:"phone_number"`
-		SendReceipt null.String `json:"send_receipt"`
-		ToFeedback  null.Bool   `json:"to_feedback"`
-		UserID      uint        `json:"user_id" binding:"required"`
-		ZipCode     null.String `json:"zip_code"`
+		Donator     models.Cardholder `json:"donator"`
+		Notes       string            `json:"notes"`
+		SendReceipt string            `json:"send_receipt"`
+		ToFeedback  bool              `json:"to_feedback"`
+		UserID      uint              `json:"user_id" binding:"required"`
 	}
 )
 
-func (cr *clientResp) BuildFromPeriodicDonationModel(pd models.PeriodicDonation) {
-	cardInfo := models.CardInfo{}
-	cardholder := models.Cardholder{}
-
-	copier.Copy(&cardInfo, &pd)
-	copier.Copy(&cardholder, &pd)
-	copier.Copy(cr, &pd)
-	cr.Cardholder = cardholder
-	cr.CardInfo = cardInfo
-	cr.PayMethod = defaultPeriodicPayMethod
+func (p *patchBody) buildPeriodicDonation() models.PeriodicDonation {
+	m := new(models.PeriodicDonation)
+	m.Cardholder = p.Donator
+	m.Notes = p.Notes
+	m.SendReceipt = p.SendReceipt
+	m.ToFeedback = p.ToFeedback
+	m.UserID = p.UserID
+	return *m
 }
 
-func (cr *clientResp) BuildFromPrimeDonationModel(pd models.PayByPrimeDonation) {
-	cardInfo := models.CardInfo{}
-	cardholder := models.Cardholder{}
+func (p *patchBody) buildPrimeDonation() models.PayByPrimeDonation {
+	m := new(models.PayByPrimeDonation)
+	m.Cardholder = p.Donator
+	m.Notes = p.Notes
+	m.SendReceipt = p.SendReceipt
+	m.UserID = p.UserID
+	return *m
+}
 
-	copier.Copy(&cardInfo, &pd)
-	copier.Copy(&cardholder, &pd)
-	copier.Copy(cr, &pd)
-	cr.Cardholder = cardholder
-	cr.CardInfo = cardInfo
+func (req clientReq) BuildTapPayReq(orderNumber string) tapPayPrimeReq {
+	primeReq := new(tapPayPrimeReq)
+	primeReq.Prime = req.Prime
+	primeReq.OrderNumber = orderNumber
+	primeReq.Amount = req.Amount
+
+	if req.Currency != "" {
+		primeReq.Currency = req.Currency
+	} else {
+		primeReq.Currency = defaultCurrency
+	}
+
+	if req.Details != "" {
+		primeReq.Details = req.Details
+	} else {
+		primeReq.Details = defaultDetails
+	}
+
+	if req.MerchantID != "" {
+		primeReq.MerchantID = req.MerchantID
+	} else {
+		primeReq.MerchantID = defaultMerchantID
+	}
+
+	primeReq.Cardholder = req.Cardholder
+	// Per required fields (even empty) of cardholder of tappay documents,
+	// use empty strings for name and phonenumber fields instead of empty.
+	if !primeReq.Cardholder.Name.Valid {
+		primeReq.Cardholder.Name = null.StringFrom("")
+	}
+
+	if !primeReq.Cardholder.PhoneNumber.Valid {
+		primeReq.Cardholder.PhoneNumber = null.StringFrom("")
+	}
+
+	primeReq.PartnerKey = globals.Conf.Donation.TapPayPartnerKey
+
+	if req.Frequency == monthlyFrequency || req.Frequency == yearlyFrequency {
+		primeReq.Remember = true
+	}
+
+	primeReq.ResultUrl = req.ResultUrl
+	return *primeReq
+}
+
+func (req clientReq) buildDraftPeriodicDonation(orderNumber string) models.PeriodicDonation {
+	m := new(models.PeriodicDonation)
+
+	m.Amount = req.Amount
+	m.Cardholder = req.Cardholder
+	m.Currency = req.Currency
+	m.UserID = req.UserID
+
+	if req.Frequency != "" {
+		m.Frequency = req.Frequency
+	} else {
+		m.Frequency = monthlyFrequency
+	}
+
+	if req.Details != "" {
+		m.Details = req.Details
+	} else {
+		m.Details = "一般線上定期定額捐款"
+	}
+
+	m.OrderNumber = orderNumber
+	m.Status = statusPaying
+
+	return *m
+}
+
+func (req clientReq) buildPrimeDraftRecord(orderNumber string, payMethod string) models.PayByPrimeDonation {
+	m := new(models.PayByPrimeDonation)
+
+	m.Amount = req.Amount
+	m.Cardholder = req.Cardholder
+	m.Currency = req.Currency
+	m.Details = req.Details
+	m.MerchantID = req.MerchantID
+	m.UserID = req.UserID
+	m.PayMethod = payMethod
+	m.OrderNumber = orderNumber
+	m.Status = statusPaying
+
+	return *m
+}
+
+func (req clientReq) buildTokenDraftRecord(orderNumber string) models.PayByCardTokenDonation {
+	m := new(models.PayByCardTokenDonation)
+
+	m.Amount = req.Amount
+	m.Currency = req.Currency
+	m.Details = req.Details
+	m.MerchantID = req.MerchantID
+	m.OrderNumber = orderNumber
+	m.Status = statusPaying
+
+	return *m
+}
+
+func (cr *clientResp) BuildFromPeriodicDonationModel(d models.PeriodicDonation) {
+	cr.Amount = d.Amount
+	cr.Cardholder = d.Cardholder
+	cr.CardInfo = d.CardInfo
+	cr.Currency = d.Currency
+	cr.Details = d.Details
+	cr.Frequency = d.Frequency
+	cr.ID = d.ID
+	cr.Notes = d.Notes
+	cr.OrderNumber = d.OrderNumber
+	cr.SendReceipt = d.SendReceipt
+	cr.ToFeedback = d.ToFeedback
+}
+
+func (cr *clientResp) BuildFromPrimeDonationModel(d models.PayByPrimeDonation) {
+	cr.Amount = d.Amount
+	cr.Cardholder = d.Cardholder
+	cr.CardInfo = d.CardInfo
+	cr.Currency = d.Currency
+	cr.Details = d.Details
+	cr.ID = d.ID
+	cr.Notes = d.Notes
+	cr.OrderNumber = d.OrderNumber
+	cr.PayMethod = d.PayMethod
+	cr.SendReceipt = d.SendReceipt
+	cr.ToFeedback = false
+	cr.Frequency = oneTimeFrequency
 }
 
 func (cr *clientResp) BuildFromOtherMethodDonationModel(d models.PayByOtherMethodDonation) {
-	copier.Copy(cr, &d)
+	cr.Amount = d.Amount
+	cr.Cardholder = models.Cardholder{
+		Name:        null.StringFrom(d.Name),
+		Email:       d.Email,
+		NationalID:  null.StringFrom(d.NationalID),
+		Address:     null.StringFrom(d.Address),
+		PhoneNumber: null.StringFrom(d.PhoneNumber),
+		ZipCode:     null.StringFrom(d.ZipCode),
+	}
+	cr.Currency = d.Currency
+	cr.Details = d.Details
+	cr.ID = d.ID
+	cr.Notes = d.Notes
+	cr.OrderNumber = d.OrderNumber
+	cr.PayMethod = d.PayMethod
+	cr.SendReceipt = d.SendReceipt
+	cr.ToFeedback = false
+	cr.Frequency = oneTimeFrequency
 }
 
 const (
@@ -163,24 +290,22 @@ const (
 	statusPaid   = "paid"
 	statusFail   = "fail"
 
-	statusPeriodicToPay  = "to_pay"
-	statusPeriodicPaying = "paying"
-	statusPeriodicPaid   = "paid"
-	statusPeriodicFail   = "fail"
-
-	primeTableName = "pay_by_prime_donations"
-
 	tapPayRespStatusSuccess = 0
 
 	defaultPeriodicPayMethod = "credit_card"
 
 	secToMsec     = 1000
 	msecToNanosec = 1000000
+
+	monthlyFrequency = "monthly"
+	yearlyFrequency  = "yearly"
+	oneTimeFrequency = "one_time"
 )
 
 // pay type Enum
 const (
-	oneTime payType = iota
+	prime payType = iota
+	token
 	periodic
 )
 
@@ -256,24 +381,42 @@ func (mc *MembershipController) CreateAPeriodicDonationOfAUser(c *gin.Context) (
 	const errWhere = "MembershipController.CreateAPeriodicDonationOfAUser"
 
 	// Validate client request
+	var err error
 	var reqBody clientReq
+	var tapPayResp tapPayTransactionResp
 
 	if failData, valid := bindRequestBody(c, &reqBody); valid == false {
 		return http.StatusBadRequest, gin.H{"status": "fail", "data": failData}, nil
 	}
 
-	userID := reqBody.UserID
+	if reqBody.Cardholder.Email == "" {
+		return http.StatusBadRequest, gin.H{"status": "fail", "data": gin.H{
+			"req.Body.donator.email": "donator email is not valid",
+		}}, nil
+	}
 
+	if reqBody.Frequency != monthlyFrequency && reqBody.Frequency != yearlyFrequency {
+		return http.StatusBadRequest, gin.H{"status": "fail", "data": gin.H{
+			"req.Body.frequency": "frequency is not supported. should be `monthly` or `yearly`",
+		}}, nil
+	}
+
+	// generate periodic donation order number
+	pdOrderNumber := generateOrderNumber(periodic, getPayMethodID(payMethodCollections[0]))
 	// Build a draft periodic donation record
-	draftPeriodicDonation := buildDraftPeriodicDonation(userID, reqBody)
+	draftPeriodicDonation := reqBody.buildDraftPeriodicDonation(pdOrderNumber)
+
+	// generate token donation order number
+	dOrderNumber := generateOrderNumber(token, getPayMethodID(payMethodCollections[0]))
+	// Build a draft card token donation record
+	draftRecord := reqBody.buildTokenDraftRecord(dOrderNumber)
 
 	// Build Tappay prime request
-	tapPayReq := buildTapPayPrimeReq(periodic, defaultPeriodicPayMethod, reqBody)
-
-	draftRecord := buildTokenDraftRecord(tapPayReq)
+	tapPayReq := reqBody.BuildTapPayReq(dOrderNumber)
+	tapPayReq.Details = fmt.Sprintf("%s;%s", "第一筆定期定額捐款", tapPayReq.Details)
 
 	// Create a draft periodic donation along with the first token donation record of that periodic donation
-	err := mc.Storage.CreateAPeriodicDonation(&draftPeriodicDonation, &draftRecord)
+	err = mc.Storage.CreateAPeriodicDonation(&draftPeriodicDonation, &draftRecord)
 	if nil != err {
 		errMsg := "Unable to create a draft periodic donation and the first card token transaction record"
 		log.Error(fmt.Sprintf("%s: %s", errWhere, errMsg))
@@ -283,16 +426,15 @@ func (mc *MembershipController) CreateAPeriodicDonationOfAUser(c *gin.Context) (
 	// Start Tappay transaction
 	tapPayReqJson, _ := json.Marshal(tapPayReq)
 
-	tapPayResp, err := serveHttp(tapPayReq.PartnerKey, tapPayReqJson)
+	tapPayResp, err = serveHttp(tapPayReq.PartnerKey, tapPayReqJson)
 
 	if nil != err {
 		if tapPayRespStatusSuccess != tapPayResp.Status {
 			// If tappay error occurs, update the transaction status to 'fail' and stop the periodic donation
-			failResp := models.PayByCardTokenDonation{
-				TappayApiStatus: null.IntFrom(tapPayResp.Status),
-				Msg:             tapPayResp.Msg,
-				Status:          statusFail,
-			}
+			failResp := models.PayByCardTokenDonation{}
+			failResp.TappayApiStatus = null.IntFrom(tapPayResp.Status)
+			failResp.Msg = tapPayResp.Msg
+			failResp.Status = statusFail
 			// Procceed even if the deletion is failed
 			mc.Storage.DeleteAPeriodicDonation(draftPeriodicDonation.ID, failResp)
 		}
@@ -302,20 +444,20 @@ func (mc *MembershipController) CreateAPeriodicDonationOfAUser(c *gin.Context) (
 		return 0, gin.H{}, models.NewAppError(errWhere, errMsg, "", http.StatusInternalServerError)
 	}
 
-	updateRecord := buildTokenSuccessRecord(tapPayResp)
-	updatePeriodicDonation := buildSuccessPeriodicDonation(tapPayResp)
-	updateRecord.ID = draftRecord.ID
+	// append tappay response onto donation model
+	tapPayResp.appendRespOnPerodicDonation(&draftPeriodicDonation)
+	tapPayResp.appendRespOnTokenDonation(&draftRecord)
 
-	if err = mc.Storage.UpdatePeriodicAndCardTokenDonationInTRX(draftPeriodicDonation.ID, updatePeriodicDonation, updateRecord); nil != err {
+	if err = mc.Storage.UpdatePeriodicAndCardTokenDonationInTRX(draftPeriodicDonation.ID, draftPeriodicDonation, draftRecord); nil != err {
 		log.Error(fmt.Sprintf("%s: %s", errWhere, err.Error()))
 	}
 
-	resp := buildClientResp(defaultPeriodicPayMethod, tapPayReq, tapPayResp)
-	resp.PeriodicID = draftPeriodicDonation.ID
-	resp.ID = draftRecord.ID
+	// build response for clients
+	resp := new(clientResp)
+	resp.BuildFromPeriodicDonationModel(draftPeriodicDonation)
 
 	// send success mail asynchronously
-	go mc.sendDonationThankYouMail(resp, "定期定額")
+	go mc.sendDonationThankYouMail(*resp, "定期定額")
 
 	return http.StatusCreated, gin.H{"status": "success", "data": resp}, nil
 }
@@ -323,31 +465,38 @@ func (mc *MembershipController) CreateAPeriodicDonationOfAUser(c *gin.Context) (
 // Handler for an authenticated user to create an one-time donation
 func (mc *MembershipController) CreateADonationOfAUser(c *gin.Context) (int, gin.H, error) {
 	const errorWhere = "MembershipController.CreateADonationOfAUser"
+	var err error
+	var reqBody clientReq
+	var tapPayResp tapPayTransactionResp
 
 	// Validate client request
 
 	// Validate pay_method
 	payMethod := c.Param("pay_method")
-	if err := validatePayMethod(payMethod); nil != err {
+	if err = validatePayMethod(payMethod); nil != err {
 		return http.StatusNotFound, gin.H{"status": "fail", "data": gin.H{"pay_method": err.Error()}}, nil
 	}
-
-	var reqBody clientReq
 
 	if failData, valid := bindRequestBody(c, &reqBody); valid == false {
 		return http.StatusBadRequest, gin.H{"status": "fail", "data": failData}, nil
 	}
 
-	userID := reqBody.UserID
+	if reqBody.Cardholder.Email == "" {
+		return http.StatusBadRequest, gin.H{"status": "fail", "data": gin.H{
+			"req.Body.donator.email": "donator email is not valid",
+		}}, nil
+	}
+
+	// generate token donation order number
+	dOrderNumber := generateOrderNumber(prime, getPayMethodID(payMethod))
+	// Build a draft card token donation record
+	draftRecord := reqBody.buildPrimeDraftRecord(dOrderNumber, payMethod)
 
 	// Start Tappay transaction
-
 	// Build Tappay pay by prime request
-	tapPayReq := buildTapPayPrimeReq(oneTime, payMethod, reqBody)
+	tapPayReq := reqBody.BuildTapPayReq(dOrderNumber)
 
-	draftRecord := buildPrimeDraftRecord(userID, payMethod, tapPayReq)
-
-	if err := mc.Storage.Create(&draftRecord); nil != err {
+	if err = mc.Storage.Create(&draftRecord); nil != err {
 		switch appErr := err.(type) {
 		case *models.AppError:
 			return 0, gin.H{}, models.NewAppError(errorWhere, "Fails to create a draft prime record", appErr.Error(), appErr.StatusCode)
@@ -358,36 +507,37 @@ func (mc *MembershipController) CreateADonationOfAUser(c *gin.Context) (int, gin
 
 	tapPayReqJson, _ := json.Marshal(tapPayReq)
 
-	tapPayResp, err := serveHttp(tapPayReq.PartnerKey, tapPayReqJson)
+	tapPayResp, err = serveHttp(tapPayReq.PartnerKey, tapPayReqJson)
 
 	if nil != err {
 		if tapPayRespStatusSuccess != tapPayResp.Status {
 			// If tappay error occurs, update the transaction status to 'fail'
+			d := models.PayByPrimeDonation{}
+			d.TappayApiStatus = null.IntFrom(tapPayResp.Status)
+			d.Msg = tapPayResp.Msg
+			d.Status = statusFail
 			mc.Storage.UpdateByConditions(map[string]interface{}{
-				"order_number": tapPayReq.OrderNumber,
-			}, models.PayByPrimeDonation{
-				TappayApiStatus: null.IntFrom(tapPayResp.Status),
-				Msg:             tapPayResp.Msg,
-				Status:          statusFail,
-			})
+				"id": draftRecord.ID,
+			}, d)
 		}
 		return 0, gin.H{}, models.NewAppError(errorWhere, err.Error(), "", http.StatusInternalServerError)
 	}
 
-	// Update the transaction status to 'paid' if transaction succeeds
-	updateRecord := buildPrimeSuccessRecord(tapPayResp)
+	// append tappay response onto donation model
+	tapPayResp.appendRespOnPrimeDonation(&draftRecord)
 
-	if err, _ := mc.Storage.UpdateByConditions(map[string]interface{}{
-		"order_number": tapPayReq.OrderNumber,
-	}, updateRecord); nil != err {
+	if err, _ = mc.Storage.UpdateByConditions(map[string]interface{}{
+		"id": draftRecord.ID,
+	}, draftRecord); nil != err {
 		log.Error(err.Error())
 	}
 
-	resp := buildClientResp(payMethod, tapPayReq, tapPayResp)
-	resp.ID = draftRecord.ID
+	// build response for clients
+	resp := new(clientResp)
+	resp.BuildFromPrimeDonationModel(draftRecord)
 
 	// send success mail asynchronously
-	go mc.sendDonationThankYouMail(resp, "單筆捐款")
+	go mc.sendDonationThankYouMail(*resp, "單筆捐款")
 
 	return http.StatusCreated, gin.H{"status": "success", "data": resp}, nil
 }
@@ -413,13 +563,9 @@ func (mc *MembershipController) PatchADonationOfAUser(c *gin.Context, donationTy
 
 	switch donationType {
 	case globals.PeriodicDonationType:
-		pd := models.PeriodicDonation{}
-		copier.Copy(&pd, &reqBody)
-		d = pd
+		d = reqBody.buildPeriodicDonation()
 	case globals.PrimeDonaitionType:
-		pd := models.PayByPrimeDonation{}
-		copier.Copy(&pd, &reqBody)
-		d = pd
+		d = reqBody.buildPrimeDonation()
 	default:
 		return http.StatusInternalServerError,
 			gin.H{"status": "error", "message": fmt.Sprintf("donation type(%s) not supported", donationType)},
@@ -520,156 +666,9 @@ func (mc *MembershipController) GetADonationOfAUser(c *gin.Context, donationType
 	return http.StatusOK, gin.H{"status": "success", "data": resp}, nil
 }
 
-func (t *tapPayPrimeReq) setDefault() {
-	t.Details = defaultDetails
-	t.Currency = defaultCurrency
-	t.MerchantID = defaultMerchantID
-}
-
-func buildClientResp(payMethod string, req tapPayPrimeReq, resp tapPayTransactionResp) clientResp {
-	c := clientResp{}
-	c.PayMethod = payMethod
-
-	c.CardInfo = resp.CardInfo
-	c.Cardholder = req.Cardholder
-	c.OrderNumber = req.OrderNumber
-
-	c.Amount = resp.Amount
-	c.Currency = resp.Currency
-	c.Details = req.Details
-
-	return c
-}
-
-func buildDraftPeriodicDonation(userID uint, req clientReq) models.PeriodicDonation {
-	m := models.PeriodicDonation{}
-
-	copier.Copy(&m, &req.Cardholder)
-
-	m.Amount = req.Amount
-	m.UserID = userID
-	m.Status = statusPeriodicPaying
-
-	return m
-}
-
-func buildPrimeDraftRecord(userID uint, payMethod string, req tapPayPrimeReq) models.PayByPrimeDonation {
-	m := models.PayByPrimeDonation{}
-
-	m.UserID = userID
-	m.PayMethod = payMethod
-
-	copier.Copy(&m, &req)
-	copier.Copy(&m, &req.Cardholder)
-
-	m.Status = statusPaying
-
-	return m
-}
-
-func buildPrimeSuccessRecord(resp tapPayTransactionResp) models.PayByPrimeDonation {
-	m := models.PayByPrimeDonation{}
-
-	copier.Copy(&m, &resp)
-	copier.Copy(&m, &resp.CardInfo)
-
-	m.TappayApiStatus = null.IntFrom(resp.Status)
-
-	ttm := time.Unix(resp.TransactionTimeMillis/secToMsec, (resp.TransactionTimeMillis%secToMsec)*msecToNanosec)
-	m.TransactionTime = null.TimeFrom(ttm)
-
-	t, err := strconv.ParseInt(resp.BankTransactionTime.StartTimeMillis, 10, strconv.IntSize)
-	if nil == err {
-		stm := time.Unix(t/secToMsec, t%secToMsec)
-		m.BankTransactionStartTime = null.TimeFrom(stm)
-	}
-
-	t, err = strconv.ParseInt(resp.BankTransactionTime.EndTimeMillis, 10, strconv.IntSize)
-	if nil == err {
-		etm := time.Unix(t/secToMsec, t%secToMsec)
-		m.BankTransactionEndTime = null.TimeFrom(etm)
-	}
-
-	m.Status = "paid"
-
-	return m
-}
-
-func buildSuccessPeriodicDonation(resp tapPayTransactionResp) models.PeriodicDonation {
-	m := models.PeriodicDonation{}
-
-	copier.Copy(&m, &resp.CardInfo)
-
-	var ciphertext string
-
-	ciphertext = encrypt(resp.CardSecret.CardToken, globals.Conf.Donation.CardSecretKey)
-	m.CardToken = ciphertext
-
-	ciphertext = encrypt(resp.CardSecret.CardKey, globals.Conf.Donation.CardSecretKey)
-	m.CardKey = ciphertext
-
-	t := time.Now()
-	m.LastSuccessAt = null.TimeFrom(t)
-	m.Status = statusPeriodicPaid
-
-	return m
-}
-
-func buildTapPayPrimeReq(pt payType, payMethod string, req clientReq) tapPayPrimeReq {
-	t := &tapPayPrimeReq{}
-
-	t.setDefault()
-
-	// Fill up required fields
-	t.Prime = req.Prime
-	t.Amount = req.Amount
-	t.Cardholder = req.Cardholder
-
-	// Per required fields (even empty) of cardholder of tappay documents,
-	// use empty strings for name and phonenumber fields instead of empty.
-	if !t.Cardholder.Name.Valid {
-		t.Cardholder.Name = null.StringFrom("")
-	}
-
-	if !t.Cardholder.PhoneNumber.Valid {
-		t.Cardholder.PhoneNumber = null.StringFrom("")
-	}
-
-	// Fill up optional fields
-	if "" != req.Currency {
-		t.Currency = req.Currency
-	}
-
-	if "" != req.Details {
-		t.Details = req.Details
-	}
-
-	if "" != req.MerchantID {
-		t.MerchantID = req.MerchantID
-	}
-
-	if (linePayResultUrl{}) != req.ResultUrl {
-		t.ResultUrl = req.ResultUrl
-	}
-
-	if "" != req.OrderNumber {
-		t.OrderNumber = req.OrderNumber
-	} else {
-		t.OrderNumber = generateOrderNumber(pt, getPayMethodID(payMethod))
-	}
-	t.PartnerKey = globals.Conf.Donation.TapPayPartnerKey
-
-	if periodic == pt {
-		t.Remember = true
-	}
-
-	return *t
-}
-
-func buildTokenSuccessRecord(resp tapPayTransactionResp) models.PayByCardTokenDonation {
-	m := models.PayByCardTokenDonation{}
-
-	copier.Copy(&m, &resp)
+func (resp tapPayTransactionResp) appendRespOnPrimeDonation(m *models.PayByPrimeDonation) {
+	m.CardInfo = resp.CardInfo
+	m.TappayResp = resp.TappayResp
 	m.TappayApiStatus = null.IntFrom(resp.Status)
 
 	ttm := time.Unix(resp.TransactionTimeMillis/secToMsec, (resp.TransactionTimeMillis%secToMsec)*msecToNanosec)
@@ -688,17 +687,42 @@ func buildTokenSuccessRecord(resp tapPayTransactionResp) models.PayByCardTokenDo
 	}
 
 	m.Status = statusPaid
-	return m
 }
 
-func buildTokenDraftRecord(req tapPayPrimeReq) models.PayByCardTokenDonation {
-	m := models.PayByCardTokenDonation{}
+func (resp tapPayTransactionResp) appendRespOnPerodicDonation(m *models.PeriodicDonation) {
+	m.CardInfo = resp.CardInfo
 
-	copier.Copy(&m, &req)
+	ciphertext := encrypt(resp.CardSecret.CardToken, globals.Conf.Donation.CardSecretKey)
+	m.CardToken = ciphertext
 
-	m.Status = statusPaying
+	ciphertext = encrypt(resp.CardSecret.CardKey, globals.Conf.Donation.CardSecretKey)
+	m.CardKey = ciphertext
 
-	return m
+	t := time.Now()
+	m.LastSuccessAt = null.TimeFrom(t)
+	m.Status = statusPaid
+}
+
+func (resp tapPayTransactionResp) appendRespOnTokenDonation(m *models.PayByCardTokenDonation) {
+	m.TappayResp = resp.TappayResp
+	m.TappayApiStatus = null.IntFrom(resp.Status)
+
+	ttm := time.Unix(resp.TransactionTimeMillis/secToMsec, (resp.TransactionTimeMillis%secToMsec)*msecToNanosec)
+	m.TransactionTime = null.TimeFrom(ttm)
+
+	t, err := strconv.ParseInt(resp.BankTransactionTime.StartTimeMillis, 10, strconv.IntSize)
+	if nil == err {
+		stm := time.Unix(t/secToMsec, t%secToMsec)
+		m.BankTransactionStartTime = null.TimeFrom(stm)
+	}
+
+	t, err = strconv.ParseInt(resp.BankTransactionTime.EndTimeMillis, 10, strconv.IntSize)
+	if nil == err {
+		etm := time.Unix(t/secToMsec, t%secToMsec)
+		m.BankTransactionEndTime = null.TimeFrom(etm)
+	}
+
+	m.Status = statusPaid
 }
 
 func createHash(data string) []byte {