+++ CHANGELOG.md
@@ -26,10 +26,23 @@
   - refactor the token generation utilties for backward compatibility.
   - enable sessions while doing google|facebook oAuth.
 
-#### New Feature: Donation endpoints
-  - /v1/users/:userID/periodic_donations endpoint with POST method
-  - /v1/users/:userID/:pay_method endpoint with POST method
+#### New Feature:
+##### Donation endpoints
+  - /v1/periodic_donations endpoint with POST method
+  - /v1/donations/:pay_method endpoint with POST method
   The above endpoints allow users to contribute monthly(the upper one) or one-time(the lower one).
+
+  - /v1/periodic_donations/:id (PATCH method)
+  - /v1/donations/prime/:id (PATCH method)
+  The above endpoints allow users to patch detailed information to the certain donation record
+
+  - /v1/periodic-donations/:id?user_id=:userID (GET method)
+  - /v1/donations/prime/:id?user_id=:userID (GET method)
+  The above endpoints allow users to get the certain donation record
+
+##### Mail endpoints
+  - /v1/mail/send_activation (POST method)
+  - /v1/mail/send_success_donation (POST method)
 
 #### Configuration refactoring
   - use `viper` to load the config